requirements.txt
appnope==0.1.0
asgiref==1.1.2
attrs==17.2.0
autobahn==17.8.1
Automat==0.6.0
channels==1.1.6
constantly==15.1.0
cycler==0.10.0
daphne==1.3.0
decorator==4.1.2
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.4
flake8==3.3.0
gunicorn==19.7.1
hyperlink==17.3.1
incremental==17.5.0
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Markdown==2.6.9
matplotlib==2.0.2
mccabe==0.6.1
numpy==1.13.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
Twisted==17.5.0
txaio==2.8.1
wcwidth==0.1.7
zope.interface==4.4.2
